move deserialization of stream configs
